requirements.txt

google-generativeai
streamlit
python-dotenv
pdfplumber
sentence-transformers
faiss-cpu
transformers
torch
pillow
numpy